move entity api dir within entity dir
