+++ frontend/system/src/Entities/api/hooks/useEntity.ts
